Crate beardog_core

BearDog Core Library

This crate provides the core functionality for the BearDog security platform.

Re-exports

pub use crate::core::BearDogCore;

pub use crate::ecosystem_simple::BearDogEcosystemProvider as SimpleBearDogEcosystemProvider;

pub use crate::ecosystem_storage::AuditConfig;

pub use crate::ecosystem_storage::EcosystemDataStore;

pub use crate::ecosystem_storage::EcosystemStorageConfig;

pub use crate::ecosystem_storage::EcosystemStorageService;

pub use crate::ecosystem_storage::RetentionPolicy;

pub use crate::ecosystem_storage::SnapshotConfig;

pub use crate::ecosystem_storage::StorageHealthReport;

pub use crate::local_optimizer::BearDogLocalOptimizer;

pub use crate::songbird_client::DiscoveredService;

pub use crate::songbird_client::RegistrationInfo;

pub use crate::universal_discovery::EcosystemCapabilityDiscovery;

pub use crate::universal_discovery::UniversalCapabilityDiscovery;

pub use crate::universal_discovery::UniversalModuleRequest;

pub use crate::universal_discovery::UniversalModuleResponse;

pub use crate::universal_optimization::EcosystemOptimizationService;

pub use crate::universal_optimization::OptimizationRequest;

pub use crate::universal_optimization::OptimizationResponse;

pub use crate::universal_optimization::PerformanceMetrics;

pub use crate::universal_optimization::UniversalOptimizationService;

pub use crate::universal_primal_provider::UniversalPrimalProvider;

pub use beardog_adapters;

pub use beardog_auth;

pub use beardog_compliance;

pub use beardog_workflows;

pub use types::*;

Modules

context_aware_licensing

Context-aware licensing - placeholder module

core

BearDog Core Module

ecosystem_simple

Simple Ecosystem Provider - HARDCODING ELIMINATED

ecosystem_storage

Ecosystem Storage Service

external_functions

Licensed External Function Implementations

licensing

local_optimizer

Local Optimization Fallback

node_registry

Node Registry implementations for BearDog Core

songbird_client

Universal Service Mesh Integration for BearDog

types

Core Types for BearDog

universal_discovery

Universal Capability Discovery System

universal_optimization

Universal Optimization Service

universal_primal_provider

Universal Primal Provider Implementation for BearDog

Structs

ServiceInfo

Enums

BearDogError

Comprehensive error types for BearDog operations

HealthStatus

Traits

BearDogService

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Type Aliases

BearDogResult

Result type alias for BearDog operations

Attribute Macros

async_trait

Derive Macros

Deserialize

Serialize